+++ app/src/main/AndroidManifest.xml
@@ -33,7 +33,7 @@
         <activity android:name=".RatingActivity" />
         <activity android:name=".DetailedViewActivity" />
         <activity android:name=".MyReviewsActivity" />
-        <activity android:name=".MentionMeActivity"></activity>
+        <activity android:name=".MentionedMeActivity"></activity>
     </application>
 
 </manifest>
