+++ app/src/main/res/menu/drawer_signed_in.xml
@@ -7,7 +7,7 @@
         />
 
     <item
-        android:id="@+id/nav_mention_me"
+        android:id="@+id/nav_mentioned_me"
         android:title="Mentioned me"
         />
